requirements.txt
# SimpleKG - Local to Global GraphRAG Implementation
# Core dependencies for the medical knowledge graph system

# Graph Database
neo4j>=5.0.0

# AWS and LLM Integration
boto3>=1.34.0
langchain-aws>=0.1.0

# Environment Management
python-dotenv>=1.0.0

# Machine Learning and NLP
scikit-learn>=1.3.0
numpy>=1.24.0

# Text Processing
nltk>=3.8.0

# Data Processing
pandas>=2.0.0
json5>=0.9.0

# Optional: For enhanced text processing
# spacy>=3.7.0
# transformers>=4.30.0

# Development and Testing (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
